requirements.txt
# Backend y API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# RAG y LLM
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.6.0
tiktoken>=0.5.0

# Vector Database
chromadb>=0.4.15
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-docx>=0.8.11
PyPDF2>=3.0.0

# Web y Frontend
jinja2>=3.1.0
aiofiles>=23.0.0

# Utilidades
python-dotenv>=1.0.0
requests>=2.31.0

# Fix para Python 3.12
setuptools>=68.0.0